Refactor: Extract useFlash tests from hooks.spec.jsx
Closes #208
