--- superglue/spec/lib/hooks.spec.jsx
@@ -1,26 +1,0 @@
-import { renderHook } from '@testing-library/react'
-import { useFlash } from '../../lib'
-import { describe, it } from 'vitest'
-import { Provider } from 'react-redux'
-import { configureStore } from '@reduxjs/toolkit'
-
-describe('hooks', () => {
-  describe('useFlash', () => {
-    it('returns the flash state', () => {
-      const preloadedState = {
-        flash: { notice: 'Hello' },
-      }
-
-      let store = configureStore({
-        preloadedState,
-        reducer: (state) => state,
-      })
-      const wrapper = ({ children }) => (
-        <Provider store={store}>{children}</Provider>
-      )
-      const { result } = renderHook(() => useFlash(), { wrapper })
-
-      expect(result.current).toEqual({ notice: 'Hello' })
-    })
-  })
-})
